feat: add script to calculate cyclometric complexity (#4)
* feat: add script to calculate cyclometric complexity

I found (and tested) a bash calculator for cyclometric complexity, and with some spot checks I think
it works fairly well - longer scripts with arrays and nested functions tend to get higher scores, and scripts with just
directives (and maybe a module load or nvidia smi) get no output (there is nothing of logic to parse) and we can
assign them a value of 0. I am generating this data now, and writing into an sqlite database with the sha256 and
sha1 so we can more easily load into our s3. When this initial set of work finishes, I will also make a plot
to show the breakdown of scores. Notably, each csv output includes the script broken into functions that
are discovered, but I decided to calculate a mean, because we only need one. I will update the PR when
the script is finished running and I have a plot, and likely run the remainder on the LC jobspecs over the next few days.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>