feat: add catalog of summary data
This adds the summarize-jobspecs.py, which will calculate
the counts for applications in the jobspecs from GitHub.
The software_tags represents results from a previous gemini
parsing, where we allowed it to identify an application but we
did not give a template. The software_tags_with_template has
the template, and is an overall cleaner (better) set. I do
think we need to do two things, 1. normalize the jobspecs by
repo (e.g., one user might have thousands for one application) and
2. possibly get our application list from elsewhere. The
data is saved to an sqlite database, and the README shows how to
read it. Finally, the data for this parsing is under scripts/data
jobspec-summary. The work also calculates the length of the jobspecs
(the best set looks to be under 10K) and the workload manager
counts. Slurm wins.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>